refactor: 빼놓은 Exception 패키지 변경
