+++ src/main/java/com/travel/role/global/auth/service/AuthService.java
@@ -30,8 +30,8 @@ import com.travel.role.global.auth.dto.TokenResponse;
 import com.travel.role.domain.user.dto.auth.LoginRequestDTO;
 import com.travel.role.domain.user.dto.auth.SignUpRequestDTO;
 import com.travel.role.global.auth.dto.TokenMapping;
-import com.travel.role.global.auth.exception.InvalidTokenException;
-import com.travel.role.global.auth.exception.NotExistTokenException;
+import com.travel.role.global.exception.auth.InvalidTokenException;
+import com.travel.role.global.exception.auth.NotExistTokenException;
 import com.travel.role.global.auth.service.mail.MailService;
 import com.travel.role.global.exception.user.AlreadyExistUserException;
 import com.travel.role.global.util.PasswordGenerator;
