+++ src/main/java/com/travel/role/global/exception/auth/TokenExceptionHandlerFilter.java
@@ -1,4 +1,4 @@
-package com.travel.role.global.auth.exception;
+package com.travel.role.global.exception.auth;
 
 import java.io.IOException;
 import java.time.LocalDateTime;
